fix api module
